move create_context to admin module
